// src/FE/src/api/generated/models/customerInfo.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CustomerInfo {
  /** @nullable */
  customerId: string | null;
  /** @nullable */
  name: string | null;
  /** @nullable */
  address: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  contactPerson: string | null;
  /** @nullable */
  phone: string | null;
}