// src/FE/src/api/generated/models/updateJobRequest.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { CreateJobObservationRequest } from './createJobObservationRequest';
import type { CreateJobWorkRequest } from './createJobWorkRequest';
import type { CustomerInfo } from './customerInfo';

export interface UpdateJobRequest {
  customer: null | CustomerInfo;
  /** @nullable */
  reportNumber: string | null;
  work: null | CreateJobWorkRequest;
  observations: null | CreateJobObservationRequest;
}